requirements.txt
fire
loguru
pyyaml
tqdm
numpy<2
pillow==9.5.0
playwright
ddddocr